feat(spring-rest): add Hibernate confi through Java file
Config file: sauja.in.api.HibernateConfig

+++ spring-rest/src/main/java/sauja/in/api/service/impl/UserServiceImpl.java
@@ -1,7 +1,10 @@
 package sauja.in.api.service.impl;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import sauja.in.api.entities.Users;
+import sauja.in.api.exception.BadRequestException;
+import sauja.in.api.exception.NotFoundException;
 import sauja.in.api.repository.UserRepository;
 import sauja.in.api.service.UserService;
 
@@ -15,27 +18,57 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Users> findAll() {
         return userRepository.findAll();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Users getUser(String id) {
-        return userRepository.getUser(id);
+        Users users=userRepository.findById(id);
+        if(users==null)
+            throw new NotFoundException("User with id "+users.getId()+" not found");
+        return users;
+    }
+
+
+
+    @Override
+    @Transactional(readOnly = true)
+    public Users findByEmail(String email) {
+        Users users=userRepository.findByEmail(email);
+        if(users==null)
+            throw new NotFoundException("User with email "+email+" not found");
+        return users;
     }
 
     @Override
+    @Transactional
     public void deleteUser(String id) {
+        Users users=userRepository.findById(id);
+        if(users==null)
+            throw new NotFoundException("User with id "+users.getId()+" not found");
         userRepository.deleteUser(id);
     }
 
     @Override
+    @Transactional
     public void modifyUser(String id, Users users) {
+        Users u=userRepository.findById(id);
+        if(u==null)
+            throw new NotFoundException("User with id "+users.getId()+" not found");
         userRepository.modifyUser(id, users);
     }
 
     @Override
+    @Transactional
     public void createUser(Users users) {
-        userRepository.createUser(users);
+        if(userRepository.findByEmail(users.getEmail())==null)
+        {
+            userRepository.createUser(users);
+            return;
+        }
+        throw new BadRequestException("User with email "+users.getEmail()+" already exists");
     }
 }
